requirements.txt
amqp==5.3.1
asgiref==3.9.1
attrs==25.3.0
billiard==4.2.1
boto3==1.40.30
botocore==1.40.30
Brotli==1.1.0
celery==5.5.3
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron_descriptor==2.0.6
cryptography==46.0.1
Django==5.2.6
django-allauth==65.11.2
django-cachalot==2.8.0
django-guardian==2.4.0
django-celery-beat==2.8.1
django-compression-middleware==0.5.0
django-cors-headers==4.8.0
django-debug-toolbar==6.0.0
django-environ==0.12.0
django-extensions==4.1
django-filter==25.1
django-health-check==3.20.0
django-jazzmin==3.0.1
django-ratelimit==4.1.0
django-redis==6.0.0
django-simple-history==3.8.0
django-storages==1.14.6
django-tenants==3.9.0
django-timezone-field==7.1
django-encrypted-model-fields==0.6.5
django_celery_results==2.6.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
factory_boy==3.3.3
Faker==37.6.0
googleapis-common-protos==1.59.1
grpcio==1.74.0
gunicorn==23.0.0
idna==3.10
importlib_metadata==8.7.0
inflection==0.5.1
iniconfig==2.1.0
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.5.4
opentelemetry-api==1.37.0
opentelemetry-distro==0.58b0
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-exporter-jaeger-proto-grpc==1.21.0
opentelemetry-exporter-jaeger-thrift==1.21.0
opentelemetry-instrumentation==0.58b0
opentelemetry-instrumentation-celery==0.58b0
opentelemetry-instrumentation-dbapi==0.58b0
opentelemetry-instrumentation-django==0.58b0
opentelemetry-instrumentation-psycopg2==0.58b0
opentelemetry-instrumentation-redis==0.58b0
opentelemetry-instrumentation-requests==0.58b0
opentelemetry-instrumentation-wsgi==0.58b0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
opentelemetry-util-http==0.58b0
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
prometheus_client==0.22.1
prompt_toolkit==3.0.52
protobuf==4.25.8
psycopg2-binary==2.9.10
pycparser==2.23
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
pytest-django==4.11.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
PyYAML==6.0.2
redis==6.4.0
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
s3transfer==0.14.0
sentry-sdk==2.37.1
six==1.17.0
sqlparse==0.5.3
thrift==0.22.0
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.10.0
wrapt==1.17.3
zipp==3.23.0
zstandard==0.24.0
